+++ src/main/java/org/moon/figura/lua/api/event/LuaEvent.java
@@ -6,8 +6,8 @@ import org.luaj.vm2.LuaFunction;
 import org.luaj.vm2.Varargs;
 import org.moon.figura.lua.LuaNotNil;
 import org.moon.figura.lua.LuaWhitelist;
-import org.moon.figura.lua.docs.LuaFunctionOverload;
 import org.moon.figura.lua.docs.LuaMetamethodDoc;
+import org.moon.figura.lua.docs.LuaMetamethodDoc.LuaMetamethodOverload;
 import org.moon.figura.lua.docs.LuaMethodDoc;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 
@@ -63,11 +63,11 @@ public class LuaEvent {
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodDoc.LuaMethodOverload(
                             argumentTypes = LuaFunction.class,
                             argumentNames = "func"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodDoc.LuaMethodOverload(
                             argumentTypes = {LuaFunction.class, String.class},
                             argumentNames = {"func", "name"}
                     )
@@ -93,7 +93,7 @@ public class LuaEvent {
 
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodDoc.LuaMethodOverload(
                     argumentTypes = String.class,
                     argumentNames = "name"
             ),
@@ -115,7 +115,7 @@ public class LuaEvent {
 
     @LuaWhitelist
     @LuaMetamethodDoc(overloads = {
-            @LuaMetamethodDoc.LuaMetamethodOverload(
+            @LuaMetamethodOverload(
                     types = {int.class, LuaEvent.class}
             )
     })
