+++ src/main/java/org/moon/figura/lua/api/nameplate/NameplateCustomizationGroup.java
@@ -2,8 +2,8 @@ package org.moon.figura.lua.api.nameplate;
 
 import org.moon.figura.lua.LuaNotNil;
 import org.moon.figura.lua.LuaWhitelist;
-import org.moon.figura.lua.docs.LuaFunctionOverload;
 import org.moon.figura.lua.docs.LuaMethodDoc;
+import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 
 import java.util.Arrays;
@@ -24,7 +24,7 @@ public class NameplateCustomizationGroup {
 
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = String.class,
                     argumentNames = "text"
             ),
