+++ src/main/java/org/moon/figura/lua/docs/FiguraGlobalsDocs.java
@@ -19,6 +19,7 @@ import org.moon.figura.lua.api.sound.SoundAPI;
 import org.moon.figura.lua.api.TextureAPI;
 import org.moon.figura.lua.api.vanilla_model.VanillaModelAPI;
 import org.moon.figura.lua.api.world.WorldAPI;
+import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
 
 /**
  * Class only exists to have docs for the global figura
@@ -79,7 +80,7 @@ public abstract class FiguraGlobalsDocs {
     public LuaFunction type;
 
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = String.class,
                     argumentNames = "scriptName"
             ),
@@ -89,12 +90,12 @@ public abstract class FiguraGlobalsDocs {
 
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload,
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload,
+                    @LuaMethodOverload(
                             argumentTypes = String.class,
                             argumentNames = "folder"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {String.class, Boolean.class},
                             argumentNames = {"folder", "subFolders"}
                     )
@@ -104,7 +105,7 @@ public abstract class FiguraGlobalsDocs {
     public static LuaTable listFiles(String folder, boolean subFolders) {return null;}
 
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = Object.class,
                     argumentNames = "arg"
             ),
@@ -113,7 +114,7 @@ public abstract class FiguraGlobalsDocs {
     public static void print() {}
 
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = Object.class,
                     argumentNames = "arg"
             ),
@@ -123,19 +124,19 @@ public abstract class FiguraGlobalsDocs {
 
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = LuaTable.class,
                             argumentNames = "table"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = LuaUserdata.class,
                             argumentNames = "javaObject"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {LuaTable.class, Integer.class},
                             argumentNames = {"table", "maxDepth"}
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {LuaUserdata.class, Integer.class},
                             argumentNames = {"javaObject", "maxDepth"}
                     )
@@ -146,19 +147,19 @@ public abstract class FiguraGlobalsDocs {
 
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = LuaTable.class,
                             argumentNames = "table"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = LuaUserdata.class,
                             argumentNames = "javaObject"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {LuaTable.class, Integer.class},
                             argumentNames = {"table", "maxDepth"}
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {LuaUserdata.class, Integer.class},
                             argumentNames = {"javaObject", "maxDepth"}
                     )
@@ -168,7 +169,7 @@ public abstract class FiguraGlobalsDocs {
     public static void logTable() {}
 
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = String.class,
                     argumentNames = "json"
             ),
@@ -177,7 +178,7 @@ public abstract class FiguraGlobalsDocs {
     public static void printJson() {}
 
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = String.class,
                     argumentNames = "json"
             ),
