--- src/main/java/org/moon/figura/lua/docs/LuaFunctionOverload.java
@@ -1,22 +1,0 @@
-package org.moon.figura.lua.docs;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface LuaFunctionOverload {
-
-    /**
-     * The types of the arguments in this overload.
-     */
-    Class<?>[] argumentTypes() default {};
-
-    /**
-     * The names to give to the arguments in this overload.
-     */
-    String[] argumentNames() default {};
-
-    Class<?> returnType() default DEFAULT.class;
-
-    final class DEFAULT {}
-}
