+++ src/main/java/org/moon/figura/lua/docs/LuaMetamethodDoc.java
@@ -9,12 +9,18 @@ import java.lang.annotation.Target;
 @Target(ElementType.METHOD)
 public @interface LuaMetamethodDoc {
 
+    /**
+     * Each {@link LuaMetamethodOverload} should indicate one set of variable types that can
+     * be used to call this method in a valid way.
+     */
     LuaMetamethodOverload[] overloads();
 
     @Retention(RetentionPolicy.RUNTIME)
     @interface LuaMetamethodOverload {
+
         //First value is the result type, rest of values are the parameter types.
         Class<?>[] types();
+
         String comment() default "";
     }
 }
