+++ src/main/java/org/moon/figura/math/vector/FiguraVec4.java
@@ -6,6 +6,8 @@ import org.luaj.vm2.LuaFunction;
 import org.moon.figura.lua.LuaNotNil;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.*;
+import org.moon.figura.lua.docs.LuaMetamethodDoc.LuaMetamethodOverload;
+import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
 import org.moon.figura.math.matrix.FiguraMat4;
 import org.moon.figura.utils.MathUtils;
 import org.moon.figura.utils.caching.CacheUtils;
@@ -37,7 +39,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec4.class
             ),
             value = "vector_n.reset"
@@ -78,11 +80,11 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec4.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w"}
                     )
@@ -113,11 +115,11 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec4.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w"}
                     )
@@ -148,11 +150,11 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec4.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w"}
                     )
@@ -170,7 +172,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = Double.class,
                     argumentNames = "factor",
                     returnType = FiguraVec4.class
@@ -201,11 +203,11 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec4.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w"}
                     )
@@ -236,11 +238,11 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec4.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w"}
                     )
@@ -271,11 +273,11 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec4.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w"}
                     )
@@ -293,7 +295,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = Double.class,
                     argumentNames = "factor",
                     returnType = FiguraVec4.class
@@ -313,7 +315,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = FiguraMat4.class,
                     argumentNames = "mat",
                     returnType = FiguraVec4.class
@@ -339,7 +341,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec4.class
             ),
             value = "vector_n.copy"
@@ -351,7 +353,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = FiguraVec4.class,
                     argumentNames = "vec"
             ),
@@ -364,7 +366,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec4.class
             ),
             value = "vector_n.normalize"
@@ -376,7 +378,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec4.class
             ),
             value = "vector_n.normalized"
@@ -388,7 +390,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = {Double.class, Double.class},
                     argumentNames = {"minLength", "maxLength"},
                     returnType = FiguraVec4.class
@@ -402,7 +404,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = {Double.class, Double.class},
                     argumentNames = {"minLength", "maxLength"},
                     returnType = FiguraVec4.class
@@ -423,7 +425,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec4.class
             ),
             value = "vector_n.to_rad"
@@ -435,7 +437,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec4.class
             ),
             value = "vector_n.to_deg"
@@ -458,7 +460,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
 
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = LuaFunction.class,
                     argumentNames = "func"
             ),
@@ -517,13 +519,13 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, FiguraVec4.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, Double.class, FiguraVec4.class}
                     )
             }
@@ -543,13 +545,13 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, FiguraVec4.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, Double.class, FiguraVec4.class}
                     )
             }
@@ -569,16 +571,16 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, FiguraVec4.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, FiguraMat4.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, Double.class, FiguraVec4.class}
                     )
             }
@@ -600,10 +602,10 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, FiguraVec4.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, Double.class}
                     )
             }
@@ -623,10 +625,10 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, FiguraVec4.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec4.class, FiguraVec4.class, Double.class}
                     )
             }
@@ -648,7 +650,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {Boolean.class, FiguraVec4.class, FiguraVec4.class}
             )
     )
@@ -658,7 +660,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {FiguraVec4.class, FiguraVec4.class}
             )
     )
@@ -668,7 +670,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {Integer.class, FiguraVec4.class}
             )
     )
@@ -678,7 +680,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {Boolean.class, FiguraVec4.class, FiguraVec4.class}
             )
     )
@@ -688,7 +690,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {Boolean.class, FiguraVec4.class, FiguraVec4.class}
             )
     )
@@ -698,7 +700,7 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {String.class, FiguraVec4.class}
             )
     )
@@ -709,13 +711,13 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {Double.class, FiguraVec4.class, Integer.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {Double.class, FiguraVec4.class, String.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVector.class, FiguraVec4.class, String.class},
                             comment = "vector_n.comments.swizzle"
                     )
@@ -754,13 +756,13 @@ public class FiguraVec4 extends FiguraVector<FiguraVec4, FiguraMat4> {
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {void.class, FiguraVec4.class, Integer.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {void.class, FiguraVec4.class, String.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {void.class, FiguraVec4.class, String.class, FiguraVector.class}
                     )
             }
