+++ src/main/java/org/moon/figura/math/vector/FiguraVec5.java
@@ -6,6 +6,8 @@ import org.luaj.vm2.LuaFunction;
 import org.moon.figura.lua.LuaNotNil;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.*;
+import org.moon.figura.lua.docs.LuaMetamethodDoc.LuaMetamethodOverload;
+import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
 import org.moon.figura.math.matrix.FiguraMatrix;
 import org.moon.figura.utils.MathUtils;
 import org.moon.figura.utils.caching.CacheUtils;
@@ -40,7 +42,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec5.class
             ),
             value = "vector_n.reset"
@@ -82,11 +84,11 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec5.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w", "t"}
                     )
@@ -118,11 +120,11 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec5.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w", "t"}
                     )
@@ -154,11 +156,11 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec5.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w", "t"}
                     )
@@ -176,7 +178,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = Double.class,
                     argumentNames = "factor",
                     returnType = FiguraVec5.class
@@ -208,11 +210,11 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec5.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w", "t"}
                     )
@@ -244,11 +246,11 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec5.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w", "t"}
                     )
@@ -280,11 +282,11 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = FiguraVec5.class,
                             argumentNames = "vec"
                     ),
-                    @LuaFunctionOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {Double.class, Double.class, Double.class, Double.class, Double.class},
                             argumentNames = {"x", "y", "z", "w", "t"}
                     )
@@ -302,7 +304,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = Double.class,
                     argumentNames = "factor",
                     returnType = FiguraVec5.class
@@ -335,7 +337,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec5.class
             ),
             value = "vector_n.copy"
@@ -347,7 +349,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = FiguraVec5.class,
                     argumentNames = "vec"
             ),
@@ -360,7 +362,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec5.class
             ),
             value = "vector_n.normalize"
@@ -372,7 +374,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec5.class
             ),
             value = "vector_n.normalized"
@@ -384,7 +386,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = {Double.class, Double.class},
                     argumentNames = {"minLength", "maxLength"},
                     returnType = FiguraVec5.class
@@ -398,7 +400,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = {Double.class, Double.class},
                     argumentNames = {"minLength", "maxLength"},
                     returnType = FiguraVec5.class
@@ -419,7 +421,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec5.class
             ),
             value = "vector_n.to_rad"
@@ -431,7 +433,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @Override
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     returnType = FiguraVec5.class
             ),
             value = "vector_n.to_deg"
@@ -454,7 +456,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
 
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = LuaFunction.class,
                     argumentNames = "func"
             ),
@@ -518,13 +520,13 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, FiguraVec5.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, Double.class, FiguraVec5.class}
                     )
             }
@@ -544,13 +546,13 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, FiguraVec5.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, Double.class, FiguraVec5.class}
                     )
             }
@@ -570,13 +572,13 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, FiguraVec5.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, Double.class, FiguraVec5.class}
                     )
             }
@@ -596,10 +598,10 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, FiguraVec5.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, Double.class}
                     )
             }
@@ -619,10 +621,10 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, FiguraVec5.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVec5.class, FiguraVec5.class, Double.class}
                     )
             }
@@ -644,7 +646,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {Boolean.class, FiguraVec5.class, FiguraVec5.class}
             )
     )
@@ -654,7 +656,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {FiguraVec5.class, FiguraVec5.class}
             )
     )
@@ -664,7 +666,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {Integer.class, FiguraVec5.class}
             )
     )
@@ -674,7 +676,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {Boolean.class, FiguraVec5.class, FiguraVec5.class}
             )
     )
@@ -684,7 +686,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {Boolean.class, FiguraVec5.class, FiguraVec5.class}
             )
     )
@@ -694,7 +696,7 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
 
     @LuaWhitelist
     @LuaMetamethodDoc(
-            overloads = @LuaMetamethodDoc.LuaMetamethodOverload(
+            overloads = @LuaMetamethodOverload(
                     types = {String.class, FiguraVec5.class}
             )
     )
@@ -705,13 +707,13 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {Double.class, FiguraVec5.class, Integer.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {Double.class, FiguraVec5.class, String.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {FiguraVector.class, FiguraVec5.class, String.class},
                             comment = "vector_n.comments.swizzle"
                     )
@@ -752,13 +754,13 @@ public class FiguraVec5 extends FiguraVector<FiguraVec5, FiguraMatrix.DummyMatri
     @LuaWhitelist
     @LuaMetamethodDoc(
             overloads = {
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {void.class, FiguraVec5.class, Integer.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {void.class, FiguraVec5.class, String.class, Double.class}
                     ),
-                    @LuaMetamethodDoc.LuaMetamethodOverload(
+                    @LuaMetamethodOverload(
                             types = {void.class, FiguraVec5.class, String.class, FiguraVector.class}
                     )
             }
